+++ _install/library/base/text/de_DE.php
@@ -7,6 +7,9 @@
 	'[[/global/adress]]' => 'Adresse',
 	'[[/global/phone]]' => 'Telefon',
 	'[[/global/email]]' => 'E-Mail',
+	
+  '[[/website/footer/title/followus]]' => 'Social Media',
+	'[[/website/footer/title/getintouch]]' => 'Hier sind wir',
 
 	// Cookie-Banner
 	'[[/cookiebanner/info/text]]' => 'Diese Website verwendet Cookies, um Ihnen die bestmögliche Erfahrung zu bieten.',