+++ _install/library/base/text/global.php
@@ -4,6 +4,10 @@
 	'[[/company/adress]]' => 'Street 1, 12345 City',
 	'[[/company/phone]]' => '+49 (0) 170 1234567',
 	'[[/company/email]]' => 'contact@example.com',
+	'[[/company/instagram]]' => 'https://www.instagram.com/your-company',
+	'[[/company/facebook]]' => 'https://www.faceboook.com/your-company',
+	'[[/company/youtube]]' => 'https://www.youtube.com/your-company',
+	'[[/company/telegram]]' => 'https://t.me/your-company',
 
 	'[[/website/url]]' => 'www.example.com',
 	'[[/website/charset]]' => 'UTF-8',