+++ _install/library/modules/navigation/text/de_DE.php
@@ -1,4 +1,3 @@
 <?php return [
-	'[[/website/footer/title/getintouch]]' => 'Hier sind wir',
 	'[[/website/footer/title/navigation]]' => 'Navigation',
 ];