+++ _install/library/modules/navigation/text/en_US.php
@@ -1,4 +1,3 @@
 <?php return [
-	'[[/website/footer/title/getintouch]]' => 'Here we are',
 	'[[/website/footer/title/navigation]]' => 'Navigation',
 ];